Fix build typings
The test sources interfered with the build as they are also covered by the tsconfig. The
build then also generated the test typings and moved the source typings into a nested
directory. As the package.json was not modified it now pointed to the parent directory
which breaks all Typescript consumers.

This is now fixed be excluding the 'test' directory from the rollup build.