Use Vuetify pickers for date and time controls
Reimplements 'date/time/date-time' controls using Vuetify pickers with the ability to
clear the data input. The input by default allows keyboard editing as well and uses
masks for keyboard entry. The pickers are shown when the associated icon is clicked.

The mask feature can be disabled using the UI Schema option 'mask=false'. The icon can be
customized using the UI Schema option 'pickerIcon=<icon-id>' with the name of the icon to
use.