feat: add touch error filtering
This commit adds a touched ref that is bound to controls from
`useVuetifyControl` and a `handleBlur` composable that updates the
`touched` state after an input is blurred.

It also takes account of a new config setting `enableFilterErrorsBeforeTouch`
which allows the forcing of all errors to display, such as when a user
attempts to submit a form which is not completely valid.